Added the 'name' and 'args' field to the 'delegate' plugin's CNI config.
When the port-mapper plugin invokes the 'delegate' plugin, the CNI
config of the 'delegate' plugin needs to have the 'name' and the
'args' fields set. We are therefore updating the config during
creation of the `PortMapper` object, and will use this config when
the 'delegate' plugin is invoked.

Review: https://reviews.apache.org/r/51740/