Made default executor handle shutdown events while disconnected.
Previously, the default executor used to crash with a failed assertion
when the executor library injected a shutdown event when it noticed
a disconnection with the agent for non-checkpointed frameworks and
upon recovery timeout for checkpointed frameworks. This change
modifies it to commit suicide minus the failed assertion.

Review: https://reviews.apache.org/r/52755/