Allow tasks to set persistent volume as readonly or readwrite resource.
Allow the task to specify the persistent volume access to be read-only
or read-write. Note that the persistent volume is always created as
read-write.

If the task is the first consumer of the shared persistent volume, then
set the ownership of the persistent volume to match that of the task.
Otherwise, launch the task but if the task is unable to read/write the
persistent volume, it may fail at that point of time.

Review: https://reviews.apache.org/r/45963/