Stout: Enabled tests that pass on Windows.
A large number of Stout test files are currently not being built on
Windows.  Many of these files contain tests for parts of Stout that have
already been ported, or require only trivial fixes to work (such as
removing `#include`s on Windows). A small minority of the tests contain
bugs that we should fix.

This commit will add these files to the build, fix some of the
trivially-fixable tests, and disable tests that are known to fail
because of bugs, including comments explaining why and links to JIRA
issues where appropriate.

Review: https://reviews.apache.org/r/50674/