Updated default value of the '--runtime_dir' agent flag.
Previously, the default value of '--runtime_dir' was hard coded to
'/var/run/mesos'. However, this directory is typically only accessable
to the 'root' user. This caused problems when launching an agent as a
non-root user. We now check to see if the agent is launched as root or
not, and change this default accordingly.

Review: https://reviews.apache.org/r/52902/