Treat failure in destroy of non-existing nested cgroups as a success.
This may happen if certain daemon on the host creates and deletes
nested cgroups inside Mesos containers but the nested cgroups are not
managed by Mesos; we shouldn't fail the destroy. Ultimately the
rationale for this change is that if the goal is to destroy a cgroup
and the end result is that it's gone, then we should consider it a
success even though some nested cgroups are cleaned up out-of-band
by other entities.

Review: https://reviews.apache.org/r/53031/