Changed master to add `FrameworkInfo` to agent reconcilation.
When an agent re-registers with a master that already knows about that
agent, the master needs to reconcile its view of the state of the agent
with the agent's current state. For any task that the master thinks
should be on the agent but isn't included in the ReregisterSlaveMessage,
the master does a reconcilation with the agent to find the task's
current state.

This commit changes adds the `FrameworkInfo` for any possibly missing
tasks to the master -> agent reconciliation message. This is useful
because the agent can consult the `FrameworkInfo` during reconciliation:
this will shortly be used to make agent reconciliation behave
differently for partition-aware frameworks.

Review: https://reviews.apache.org/r/52722/