Changed agent to send TASK_DROPPED during reconciliation.
If a framework attempts to launch a task but the launch message is
dropped after it reaches the master but before it reaches the slave, the
failed launch will be detected during master <-> agent reconciliation
when the agent re-registers. Previously, the agent would generate a
TASK_LOST status update for such dropped tasks; now it will generate
TASK_DROPPED if the framework is partition-aware.

Note that we'll only send TASK_DROPPED if the agent is running a
sufficiently recent version of Mesos (>= 1.1.0). That means that in a
mixed cluster where the master has been upgraded to Mesos 1.1 but some
of the agents have not been, a partition-aware framework might still see
TASK_LOST in this situation.

Review: https://reviews.apache.org/r/52723/