Fixed incorrect check in dynamic_cast.
We checked here whether the passed pointer is not null, while we in
fact wanted to confirm that the performed dynamic_cast succeeded
(like is already done e.g., in the definitions of flags.stringify or
flags.validate).

This change makes us check the result of the dynamic_cast like
originally intended.

Review: https://reviews.apache.org/r/53055/