Added column for FrameworkID when displaying tasks in the WebUI.
This adds some disambiguation on the Mesos Web UI front page,
which used to only show the TaskID and name.  This meant that, if
two tasks had the same ID and name, the user could only disambiguate
by digging into the tasks' sandboxes.

With the addition of a FrameworkID column, the TaskID column has
been renamed for clarity.

Review: https://reviews.apache.org/r/53324/