Fixed the container status for nested containers.
Prior to this patch, the ContainerStatus inside the TaskStatus for a
task in the default executor is not correct. The agent always gets the
container status of the top level container, even if the task might
correspond to a nested container.

This patch fixed this issue by letting the executor set the container
ID for the task (because only it has the knowledge about this
mapping), and the agent will use that container ID to get the proper
container status.

Review: https://reviews.apache.org/r/53465