Allow CREATE of shared persistent volumes based on capability.
When a framework issues a CREATE for a shared volume, we allow that
operation only if the framework has opted in for the capability of
SHARED_RESOURCES. However, for HTTP operator (/create-volumes), we
do not check as the operator API is not related to a specific
framework.

This is because it's counterintuitive if a framework can create a
shared persistent volume which cannot be offered to it.

Review: https://reviews.apache.org/r/52587/