Avoided memory leaks in exception/error paths in tests.
`clang-tidy` rightly points out that certain `ASSERT` failures can
result in leaking some heap-allocated values that aren't wrapped in a
smart pointer. This commit fixes the cases that `clang-tidy` complains
about by wrapping the values in `Owned<T>`.

Note that there are many other places in the tests that leak resources
if an exception occurs. The proper fix is usually to use a smart pointer
rather than a raw pointer; several of these APIs should be changed to
return Owned<T>, for example. However, this is not always easy/clean, in
part because of limitations of the current `Owned<T>` and `Shared<T>`
types (e.g., MESOS-5122, MESOS-6496). So for now, just fix the cases
that clang-tidy complains about and a few other similar instances.

Review: https://reviews.apache.org/r/53387/