Fixed `DefaultExecutorTest.*` tests on hosts without Docker.
Some of the tests are parameterized by the containerizers.

Fixed the naming schema, so that the tests using the Docker
containerizer run only if Docker is available.

Now the tests with the Mesos containerizer also run if the tests
are started as a non-root user.

Review: https://reviews.apache.org/r/53266/