Windows: Fixed provisioner copy backend build error.
The provisioner's `copy` backend has never been supported on Windows,
but we have included it in the build to reduce the number of `#ifdef`s
needed to compile the agent.  If the user attempts to specify a rootfs
on Windows, it will still fail in the containerizer's validation logic.

Review: https://reviews.apache.org/r/53623