Added a default PATH environment variable when launching a container.
Having an environment set but no PATH variable set inside of it,
will trigger an `execvpe` call in launch.cpp on the command we are
trying to launch. Normally this doesn't casue problems because
`execvpe` with call `confstr(_CS_PATH)` under the hood to set up a
default path to use when lookiing up our command.

However, when provisioning a new filesystem for a container, its
possible that the `confstr(_CS_PATH)` called from the agent may not
return the same path as a `confstr(_CS_PATH)` call would return using
the libc installed inside the container. This can lead to problems
(for example) with finding the `sh` command in containers based on an
alpine linux image.

We observed this in our test setup with `confstr(_CS_PATH)` on the
agent returning `/usr/bin`, while alpine linux only has the `sh`
command installed in `/bin`.

Review: https://reviews.apache.org/r/53585/