Improved performance for `getQuotaRoleAllocatedResources` in allocator.
In the allocator, `getQuotaRoleAllocatedResources` is trying to omit
dynamic and static  reservations, and persistent volumes. But it is
traversing all `Resource` objects from the `Resources` object, strips
the `role` and then using `+=` to add each `Resource` object back.

The `Resources& +=(const Resource&)` operation performs validation of
the `Resource` object which will impact the performance negatively.
This fix is calling `flatten()` directly to strip the `role` for all
of the `Resource` objects and this will not perform validation.

Review: https://reviews.apache.org/r/53929/