Added missing cleanup to libprocess 'finalize()'.
The `finalize()` function in libprocess previously did
not delete the thread-local `Executor` which is used to
defer execution to an unspecified context. This object
must be deleted during finalization so that the
`__executor__` macro creates a new process if libprocess
is subsequently reinitialized (and to not leak the pointer).

Review: https://reviews.apache.org/r/53817/