Fix SSL downgrade pathway for temporary/persistent sockets.
This fixes some potential CHECK failures when a libprocess process
has (1) SSL downgrade enabled and (2) temporary and persistent
connections open with the same remote address.  The second point is
only possible if messages are to a remote address without a persistent
connection and then a persistent connection is created.

The SSL downgrade path was only checking if the address of a socket
matched when performing the downgrade.  The code must also check to
see if the socket itself matches.

Review: https://reviews.apache.org/r/53997/