Made MesosContainerizer launch helper to take ContainerLaunchInfo.
Prior to this patch, the launch helper takes various flags from
MesosContainerizer to launch the container. This makes it very hard to
add more parameters to the launch helper. This patch simplifies this
by passing 'ContainerLaunchInfo' instead. 'ContainerLaunchInfo' is
also the protobuf message returned by isolators during 'prepare()'.
This makes it very easy to merge them and send it to the launch
helper. More importantly, this makes it very easy to add more
parameters to the launch helper in the future.

Review: https://reviews.apache.org/r/54103