CMake: Changed example module output location and depedencies.
This moves some test libraries from the default CMake build
output locations to `<build_dir>/src/.libs`, which is where
the Automake build outputs libraries.  We have a few tests
that expect modules to be located in this directory.

Additionally, this makes the test target (make check) depend
on the test modules and example frameworks, so that they are built
in the correct order.

Review: https://reviews.apache.org/r/53760/