CMake: Add a target between MESOS_TARGET and MESOS_PROTOBUFs.
We need this extra level of targets in order to model the
dependency chain for binaries and libraries.  An example
is the mesos-containerizer binary, which depends on libmesos
(i.e. MESOS_TARGET) but does not get included in the dependency
chain if you only depend on MESOS_TARGET.

Review: https://reviews.apache.org/r/53761/