Added helper to get the io switchboard server address.
For now, the io switchboard server will always hosted on a unix domain
socket. The path limit for a unix domain socket is 100 bytes, so we
can't just use a file in our runtime directory as the socket file we
connect to. Instead, we store the path to the socket file in our
runtime directory and read it from there.

Review: https://reviews.apache.org/r/53938/