Avoid empty `Resources` entries in `Master::Slave::usedResources`.
When removing the last task from the agent struct in the master,
we ensure that we do not leave an empty `Resources` entry for its
framework. When removing the last executor, on the other hand, we
may leave an empty `Resources`! This patch ensures that we always
clean up `usedResources[frameworkId]` if the `Resources` are empty.

This patch also simplifies the condition to more directly look
for empty resources (previously this was inferred from there
being no tasks or executors remaining).

Along the way, replace using `[]` to access a hashmap in a `CHECK` with
`at`. Because `[]` can modify the map, we should generally prefer using
`at` for accesses that are intended to be read-only.

Review: https://reviews.apache.org/r/54223/