Windows: Added support for Agent test harness.
Several files required to run any of the Agent tests currently can't be
compiled by MSVC. The individual reasons are mostly small, incidental
details -- missing headers, assumption of POSIX paths, and so on. The
one big exception to this is that the `main` function of the test binary
itself isn't setting up the WSA stack and making sure the right SEH
callbacks are present.

This commit addresses all of these issues.

Review: https://reviews.apache.org/r/52778/