Fixed switchboard build break on Windows.
This commit will fix a build break on Windows 10, MSVC 1900, introduced
dually by commits `bd1bd1eeee50dcb23d65dcadb4903822ffd88cfb` and
`071ed5226be790aecd5ace1a7abda30b682078be`.

The recent additon of the I/O switchboard includes two things that break
the Windows build. This commit will correct both of them.

First, `switchboard_main.cpp` references several constructs that are
conditionally compiled out of the Windows build. One example is
`IOSwitchboardServerFlags`. This was never supposed to be built on
Windows, but it was accidentally added to the build, so this commit will
simply remove it from the Windows build.

Second, the I/O switchboard changes introduced a new flag,
`--io_switchboard_enable_server`, which is not implemented on Windows.
When this change was submitted, this flag was accidentally conditionally
not compiled on Windows builds, even though it is referenced in
`switchboard.cpp` (since initialization of the switchboard is different
when the server is not turned on). This commit will fix this issue by
simply moving the `#ifdef` down so that the symbol is defined and used
properly.

Review: https://reviews.apache.org/r/54324/