Fixed HttpServeTests to connect to a non-ANY address.
Connecting to the ANY address (0.0.0.0) is legal in BSD sockets,
but results in an ambiguous connection.  The outgoing connection
could literally be anywhere on the network.

These tests aim to connect to a specific socket on the local
machine, and as such, should supply a valid IP address.
This commit coincidentally fixes these tests on Windows,
which disallows connecting to invalid IP addresses.

Review: https://reviews.apache.org/r/54328