Windows: Fixed agent `--runtime_dir` flag default value.
This commit fixes MESOS-6677, which breaks the ability
to run any agent on Windows, and thus is blocking all
Windows development progress on the `master` branch.

The cause was that the default `runtime_dir` value was POSIX specific,
and used `os::user()` which is not implemented on Windows.
The fix is to guard the POSIX code and add a Windows implementation.

Review: https://reviews.apache.org/r/54336/