Windows: Added `authentication_tests.cpp` to the build.
This commit will partially address MESOS-6697 by adding
`authentication_tests.cpp` to the Windows build, but disabling all
tests in the file, as the Windows implementation of the Master currently
does not support authentication (see MESOS-6733).

Our goal in adding tests to the build is to ensure that the tests are
compilable on Windows as we continue to work on MESOS-6733. When this
issue is resolved, we do expect to come back and enable these tests,
which should resolve MESOS-6697.

Review: https://reviews.apache.org/r/54446/