Remove use of `typename` causing Windows build break.
The use of `typename` in C++ is used to disambiguate between static
members and dependent names that use the same symbol. For example, if
you have some `T::iterator` and `T` happens to have a static member
called iterator, it is necessary to add `typename T::iterator` to
indicate that you want an iterator of `T` rather than to refer to teh
static member `T::iterator`.

In some cases we employ `typename` to make our intention clearer, even
though it is not strictly speaking necessary. While normally a good
habit, in the specific cases we change in this review, it causes MSVC to
explode.

This commit will remove these uses.

Review: https://reviews.apache.org/r/53552/