Made the IO Switchboard wait for the reader to signal before exiting.
The switchboard never used to wait for the read end of the pipe
to finish processing the data before exiting. This was problematic
as the client connected to the output entrypoint might not receive
the data. Note that we are relying on `socket.close()` to
eventually ensure that the data is flushed to the client when the
socket is closed.

Review: https://reviews.apache.org/r/54496/