Refactored the reaping logic in I/O switchboard.
Previously, we don't handle the case where reaping failed. This patch
refactored the reaping logic to address the issue. Here are the
changes in this patch:

(1) Add a reaped() function which we call whenever a process is
    reaped. From there we log why the process was reaped and
    (possibly) set a limitation on it if it died abnormally.
    We store this limitation in the switchboard's info struct.

(2) Simplify the logic in watch() down to do nothing more than return
    the limitation future from the switchboard's info struct.

(3) Change cleanup() to only send SIGTERM to the switchboard process
    if its status is still pending (i.e. it hasn't been reaped yet).

(4) Change cleanup to properly remove the unix domain socket file in
    all cases of the reaped() future being satisifed.

Review: https://reviews.apache.org/r/54518