Fixed inaccurate documentation of `--hostname` for master and agent.
This comment is inaccurate since the introduction of
`LIBPROCESS_ADVERTISE_IP`, which introduced the distinction between the
IP that we bind to vs advertise.

If the `--hostname` flag is unset, it is resolved from the IP address
that the master/agent advertises. This may happen to be same as the IP
address that we bind to, but ultimately we resolve it based on the
advertise IP.

Review: https://reviews.apache.org/r/54275/