Added a grace period for I/O switchboard server termination.
We give the I/O switchboard server a grace period to wait for the
connection from the containerizer. This is for the case where the
container itself is short lived (e.g., a DEBUG container does an 'ls'
and exits). For that case, we still want the subsequent attach output
call to get the output from that container.

Review: https://reviews.apache.org/r/54562