Stout: Added logic for TMPDIR environment variable in `os::temp`.
`TMPDIR` is a POSIX-standard environment variable which can be used
to specify a temporary directory.  This variable is currently read
in the agent tests, but ignored in other parts of the codebase.
(`os::temp` is commonly used by `os::mkdtemp`.)

This commit is one of two commits that will normalize the location
of the temporary directory.

Review: https://reviews.apache.org/r/54489/