Updated `Role::resources()` in master to handle multi-role frameworks.
Currently, the `Role::resources()` method in the `master::Role` struct
returns the total resources used and offered to a framework, which will
be incorrect for multi-role frameworks.

For now, since the support for multi-role frameworks is incomplete
(in particular, the `Resource.allocation_info` is not yet populated),
we continue to handle single-role frameworks in the old manner. Once
the `Resource.allocation_info` is populated, we will remove the
single-role special case logic in `Role::resources()`.

Review: https://reviews.apache.org/r/54524/