Added SSL support to libprocess HTTP request helpers.
In order to test libprocess HTTP functionality with SSL
both enabled and disabled, the HTTP helper functions that
libprocess uses in tests must be parametrized by SSL
configuration.

The HTTP/HTTPS scheme appears as the last argument in these helpers
because it is the least commonly changed part of the URL.
Most tests that do not exercise SSL will suffice with a default.

Review: https://reviews.apache.org/r/50736/