Changed hostname used for SSL cert creation in tests.
This changes the SSL helpers in libprocess to generate
certs using a hostname determined using the same address
that is advertised by libprocess. This assures that
validation of the certificates will succeed, even in environments
where the loopback interface does not resolve to a single hostname.

The test `SSLTest.BasicSameProcess` is also updated to accommodate
this change.

Review: https://reviews.apache.org/r/51065/